+++ src/calls/feedback.js
@@ -1,0 +1,27 @@
+import apiCall from "../api-call";
+
+export default object =>
+  Object.assign(object, {
+    createFeedback: apiCall({
+      method: "post",
+      path: "/feedbacks",
+      expectedParams: ["body"],
+      optionalParams: ["anonymous"]
+    }),
+
+    getFeedback: apiCall({
+      method: "get",
+      path: "/feedbacks/:feedbackId"
+    }),
+
+    listFeedback: apiCall({
+      method: "get",
+      path: "/feedbacks",
+      optionalParams: ["page"]
+    }),
+
+    reviewFeedback: apiCall({
+      method: "patch",
+      path: "/feedbacks/:feedbackId"
+    })
+  });