+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.7] - 2017-10-02
 ### Added
 - Add requestPasswordReset and resetPassword commands for a "Forgot Password" flow.
 - Add the comment CRUD actions.
@@ -45,13 +48,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Allow GET requests to have query parameters.
 - Mark responses with 3xx status codes as successful.
 
-## 0.0.1 - 2017-09-22
+## [0.0.1] - 2017-09-22
 ### Added
 - Initial checkin.
 
-[Unreleased]: https://github.com/CultureHQ/client/compare/v0.0.6...HEAD
+[Unreleased]: https://github.com/CultureHQ/client/compare/v0.0.7...HEAD
+[0.0.7]: https://github.com/CultureHQ/client/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/CultureHQ/client/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/CultureHQ/client/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/CultureHQ/client/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/CultureHQ/client/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/CultureHQ/client/compare/v0.0.1...v0.0.2
+[0.0.1]: https://github.com/CultureHQ/client/compare/f26481...v0.0.1